chore: Split figure markets democratized prime project into sibling folders for each demo prime pool
